Split the generator in multiple files for refactoring.
Related to #76.
